# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Laravel
  description: ''
  version: 1.0.0
servers:
  -
    url: 'http://localhost'
paths:
...
  /api/login-admin:
    post:
      summary: loginAdmin
      operationId: loginAdmin
      description: ''
      parameters: []
      responses:
        200:
          description: ''
          content:
            text/plain:
              schema:
                type: string
                example: "{\n   token:\"8|8MxQJgQICxUG4L7UUbS91dovFpjGhZj8dDJA17c8\"\n}"
        400:
          description: ''
          content:
            text/plain:
              schema:
                type: string
                example: "scenario = \"user send a valid token\"\n\"message\": \"You are already logged in\""
        401:
          description: ''
          content:
            text/plain:
              schema:
                type: string
                example: "scenario = \"Email or Password Wrong\"{\n\"Message\": \"Email or Password Wrong\"\n}"
        422:
          description: ''
          content:
            text/plain:
              schema:
                type: string
                example: "scenario= \"validation errors\"{\n\"message\": \"The email field is required. (and 1 more error)\",\n\"errors\": {\n\"email\": [\n\"The email field is required.\"\n],\n\"password\": [\n\"The password field is required.\"\n]\n}\n}"
      tags:
        - Endpoints
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  description: 'Must be a valid email address.'
                  example: jillian94@example.org
                password:
                  type: string
                  description: ''
                  example: necessitatibus
              required:
                - email
                - password
      security: []
...